+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Other:            gnss: simulation considers more apriori models (e.g. TEC maps).
 - Other:            IGRF: Updated International Geomagnetic Reference Field (IGRF) to 14th Generation Release
 - Other:            GNSS: Improved setup of ambiguity parameters. Considers splitted network, splitted observations (e.g. L2LG, L2WG).
+- Other:            GNSS: More stable estimation of ambiguities. Faster ambiguity resolution (decorrelation step).
 - Other:            Store more double digits in xml format.
 - Other:            NormalsEliminate: regularize unused eliminated parameters.
 - Other:            Sinex2StationPosition: Apply frequencies to all intervals.